Updated the release pipeline
Created a script to manage the releases to RubyGems.
Created a script to read the version of the gem from the version.rb file. Both of these newly created scripts are used in the release workflow.
Added support for creating Github releases for Forsis.
Added Gemfile.lock to version control.